Module Tjr_impfs_v1.V1_generic

An abstract generic development; this layer deals with path resolution, times and a few other things; assumes create_dir and create_file.

type exn_ = Tjr_minifs.Minifs_intf.exn_

type stat_times = Tjr_minifs.Minifs_intf.times

type stat_record = Tjr_minifs.Minifs_intf.stat_record

module type S0 = sig ... end

Very basic types

module S1 : functor (S0 : S0) -> sig ... end

Types created from basic types

module S2 : functor (S0 : S0) -> sig ... end

The values we expect to be present

module type T = sig ... end

The resulting filesystem (fd is a file identifier, dh is a directory handle supporting leaf_stream operations)

module Make : functor (S0 : S0) -> sig ... end

module X = Make