Module Tjr_impfs_v1__V1

V1 - implement directories and file metadata, with file data passed through to an underlying filesystem.

We have a single "global object map", which maps ids (ints) to either dir-roots, or file-meta, or symlinks.

NOTE much of this is based on Tjr_minifs.In_mem, but with on-disk maps, rather than in-memory.

module G = Tjr_impfs_v1.Generic

type ('k, 'v, 't) uncached_btree = ('k, 'v, 't) Tjr_btree.Make_3.uncached_btree

val buf_create : unit -> Tjr_fs_shared__.Buf_factory.Buf_as_bigarray.ba_buf

val s256_to_string : Tjr_fs_shared.str_256 -> string

Setup the generic instance

module S0 : sig ... end

Base types

module Make : sig ... end

module S1 = Make.S1

derived types

module S2 = Make.S2

functor argument type holder

module type T2 = S2.T2

what we need to provide to get a filesystem

module Make_2 = Make.Make_2

functor to construct a filesystem (takes a module of type T2); the implementation is T2_impl, towards the end of this file

Some simple defns we can complete here

val root_did : S0.did

val mk_stat_times : unit -> (G.stat_times, Tjr_monad.With_lwt.lwt) Tjr_monad.m

Implement definitions from T2, ie, the core filesystem

type obj_root = Tjr_fs_shared.Std_types.blk_id

val bin_shape_obj_root : Bin_prot.Shape.t

val bin_size_obj_root : obj_root Bin_prot.Size.sizer

val bin_write_obj_root : obj_root Bin_prot.Write.writer

val bin_writer_obj_root : obj_root Bin_prot.Type_class.writer

val __bin_read_obj_root__ : (int -> obj_root) Bin_prot.Read.reader

val bin_read_obj_root : obj_root Bin_prot.Read.reader

val bin_reader_obj_root : obj_root Bin_prot.Type_class.reader

val bin_obj_root : obj_root Bin_prot.Type_class.t

type dir_entry = S1.dir_entry

Various marshallers

module Ms : sig ... end

Various runtime components that get filled in later

val blk_dev_ref : (Tjr_fs_shared.Std_types.blk_id, Tjr_fs_shared.Std_types.blk, Tjr_fs_shared.Std_types.t) Tjr_fs_shared.blk_dev_ops option Stdlib.ref

val blk_dev_ops : (Tjr_fs_shared.Std_types.blk_id, Tjr_fs_shared.Std_types.blk, Tjr_fs_shared.Std_types.t) Tjr_fs_shared.blk_dev_ops Stdlib.Lazy.t

val blk_alloc_ref : (Tjr_fs_shared.Std_types.blk_id, Tjr_fs_shared.Std_types.t) Tjr_fs_shared.blk_allocator_ops option Stdlib.ref

val blk_alloc : (Tjr_fs_shared.Std_types.blk_id, Tjr_fs_shared.Std_types.t) Tjr_fs_shared.blk_allocator_ops Stdlib.Lazy.t

val root_ops_ref : (Tjr_fs_shared.Std_types.blk_id, Tjr_fs_shared.Std_types.t) Tjr_monad.with_state option Stdlib.ref

val root_ops : (Tjr_fs_shared.Std_types.blk_id, Tjr_fs_shared.Std_types.t) Tjr_monad.with_state Stdlib.Lazy.t

val min_free_id_ref : int Stdlib.ref

We use mutually exclusive subsets of int for identifiers

val new_id : unit -> (int, Tjr_monad.With_lwt.lwt) Tjr_monad.m

val new_did : unit -> (S0.did, Tjr_monad.With_lwt.lwt) Tjr_monad.m

val new_fid : unit -> (S0.fid, Tjr_monad.With_lwt.lwt) Tjr_monad.m

val new_sid : unit -> (S0.sid, Tjr_monad.With_lwt.lwt) Tjr_monad.m

The global object map (GOM)

module Gom : sig ... end

val gom_btree : (Gom.k, Gom.v, S0.t) uncached_btree Stdlib.Lazy.t

Definitions after the GOM has been initialized

module With_gom : functor () sig ... end

NOTE Instantiate this after gom_btree has been initialized